requirements.txt
tensorflow==2.16.1
numpy==1.26.4
pillow==10.0.0
Flask==3.0.0